+++ src/KingmakerGunslinger/Blueprints/GunslingerClassBlueprints.cs
@@ -24,7 +24,8 @@ namespace KingmakerGunslinger.Blueprints
             PistolWhipBlueprintSet pistolWhip, UtilityShotBlueprintSet utilityShot,
             GunTrainingBlueprintSet gunTraining, DeadShotBlueprintSet deadShot,
             StartlingShotBlueprintSet startlingShot,
-            TargetingHeadBlueprintSet targetingHead)
+            TargetingHeadBlueprintSet targetingHead,
+            TargetingTorsoBlueprintSet targetingTorso)
         {
             CharacterClass = characterClass ?? throw new ArgumentNullException("characterClass");
             Progression = progression ?? throw new ArgumentNullException("progression");
@@ -43,6 +44,8 @@ namespace KingmakerGunslinger.Blueprints
                 throw new ArgumentNullException("startlingShot");
             TargetingHead = targetingHead ??
                 throw new ArgumentNullException("targetingHead");
+            TargetingTorso = targetingTorso ??
+                throw new ArgumentNullException("targetingTorso");
         }
         internal BlueprintCharacterClass CharacterClass { get; private set; }
         internal BlueprintProgression Progression { get; private set; }
@@ -59,7 +62,8 @@ namespace KingmakerGunslinger.Blueprints
         internal DeadShotBlueprintSet DeadShot { get; private set; }
         internal StartlingShotBlueprintSet StartlingShot { get; private set; }
         internal TargetingHeadBlueprintSet TargetingHead { get; private set; }
-        internal int Count { get { return 4 + Grit.Count + Deadeye.Count + Dodge.Count + QuickClear.Count + Nimble.Count + PistolWhip.Count + UtilityShot.Count + GunTraining.Count + DeadShot.Count + StartlingShot.Count + TargetingHead.Count; } }
+        internal TargetingTorsoBlueprintSet TargetingTorso { get; private set; }
+        internal int Count { get { return 4 + Grit.Count + Deadeye.Count + Dodge.Count + QuickClear.Count + Nimble.Count + PistolWhip.Count + UtilityShot.Count + GunTraining.Count + DeadShot.Count + StartlingShot.Count + TargetingHead.Count + TargetingTorso.Count; } }
     }
 
     internal sealed class GunslingerClassCatalogPublication
@@ -155,6 +159,8 @@ namespace KingmakerGunslinger.Blueprints
                 StartlingShotBlueprints.Register(registry);
             TargetingHeadBlueprintSet targetingHead =
                 TargetingHeadBlueprints.Register(registry);
+            TargetingTorsoBlueprintSet targetingTorso =
+                TargetingTorsoBlueprints.Register(registry);
             BlueprintProgression progression = registry.Register<BlueprintProgression>(
                 ProgressionSymbol, () => CreateProgression());
 
@@ -172,13 +178,14 @@ namespace KingmakerGunslinger.Blueprints
             progression.LevelEntries[6].Features.Add(deadShot.Feature);
             progression.LevelEntries[6].Features.Add(startlingShot.Feature);
             progression.LevelEntries[6].Features.Add(targetingHead.Feature);
+            progression.LevelEntries[6].Features.Add(targetingTorso.Feature);
             Validate(characterClass, progression, proficiencies, fullBab, goodSave,
                 poorSave, startingPistol, blackPowder, leadBall,
                 simple, martial, lightArmor, firearmProficiency);
             return new GunslingerClassBlueprintSet(characterClass, progression,
                 proficiencies, grit, deadeye, dodge, quickClear, nimble, initiative,
                 pistolWhip, utilityShot, gunTraining, deadShot, startlingShot,
-                targetingHead);
+                targetingHead, targetingTorso);
         }
 
         internal static GunslingerClassCatalogPublication Publish(